+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * PHP 7.3.4 is now the minimum required version.
+* `Duration` now extends `DateInterval` and can be used interchangeably.
+* Objects having a `__toString()` method are now supported values.
 
 ## 4.0 - 2019-09-03